requirements-pypi.txt
absl-py==2.1.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
bitsandbytes==0.43.0
braceexpand==0.1.7
brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses==0.6
datasets==2.18.0
decord==0.6.0
diffusers==0.27.0
dill==0.3.8
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
exifread-nocycle==3.0.1
ffmpeg-python==0.2.0
filelock==3.13.1
fire
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==1.0.0
gitdb==4.0.11
gitpython==3.1.42
grpcio==1.62.0
hf-transfer==0.1.5
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
img2dataset==1.45.0
importlib-metadata==7.0.1
jinja2==3.1.3
joblib==1.3.2
kornia==0.7.1
langdetect==1.0.9
lazy-loader==0.3
markdown==3.5.2
markupsafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mutagen==1.47.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.1
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodomex==3.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scenedetect==0.6.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sentry-sdk==1.41.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
timeout-decorator==0.5.0
tokenizers==0.15.2
torch==2.2.1
torchdata==0.7.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.37.1
triton==2.2.0
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
video2dataset
wandb==0.16.4
wcwidth==0.2.13
webdataset==0.2.86
websockets==12.0
webvtt-py==0.4.6
werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
yt-dlp==2023.12.30
zipp==3.17.0

